Add CircleMenuForUI with ImageBitmap. Remove userImages from Coil. Migrate from AsyncImage to Image with CircleMenuForUI in CircleMenu UIs

+++ app/src/main/java/com/kindeev/swipelauncher/presentation/activities/MainActivity.kt
@@ -102,6 +102,7 @@ class MainActivity : ComponentActivity() {
         CoroutineScope(Dispatchers.IO).launch {
             launch {
                 container.circleMenus.collect { allCircleMenus ->
+                    if (allCircleMenus.isEmpty()) return@collect
                     container.userImagesRepository.removeUnused(
                         allCircleMenus.getUsedImagesIds().toSet()
                     )
